Use working block to determine multicast start
Introduces another possible state for the working block - when a node
has received a BlockNew, but hasn't started the multicast protocol yet.
Additionally introduces further abstraction for dealing with multicast
messages. Changes where checkpoints are started, so that they don't
start while the multicast protocol is in action. Also checks to make
sure that the latest BlockNew will be built on top of the current chain
head.

Signed-off-by: Bridger Herman <bridger@bitwise.io>