Remove extra services from CFT compose file
The CFT docker compose tests file had extra services defined that had no
command to run on start, meaning they exited immediately and ended the
tests. They were there to ensure that the services existed for the
tests. It's possible to use `--no-abort` to prevent this behavior, but
if we move these services out to a separate file we won't have to use
that flag, and the separate compose file can be used across multiple
different tests.

Signed-off-by: Kenneth Koski <knkski@bitwise.io>