Provide a more robust catchup implementation
Able to catch up to the rest of the network after having crashed without
any state saved.

Allows primary to catch up, as it may behind after a view change to a
node that's still catching up.

Converts the received BlockNew message to be a Commit message before
sending it off to `handlers::commit` for processing as a commit message.

Ensures that the BlockNew message we receive gets the right sequence
number.

Ensures that `state.seq_num` is synced after doing all of the catchup
processing.

Rejects blocks that we've seen before outright.

Signed-off-by: Kenneth Koski <knkski@bitwise.io>