Check view number matches when adding msg to log
Updates `add_message` to check all messages that aren't Checkpoints or
ViewChanges to see if the view number matches the node's current view;
if it does not, do not add the message to the log. This is guarantee of
the PBFT algorithm that was not previously implemented. Also requires
passing state to `add_message`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>