Get working block in `on_block_valid`
When sending a Commit message in `on_block_valid`, we should use the
working block instead of getting the block from the consensus service
for two reasons:
- It is more efficient and less error-prone since it doesn't require a
call to the validator; we can simply get it from state.
- The validator's handler for `get_blocks` is broken in that it does not
include the `summary` for the blocks it returns; we will need the
summary of the block to make sure that the block in the Commit message
matches the block in the corresponding PrePrepare message.

Signed-off-by: Logan Seeley <seeley@bitwise.io>