Improve BlockNew handling and PrePrepare checking
Updates how BlockNew and PrePrepare messages are handled and checked in
the following ways:
- When a BlockNew is received, set its sequence number to the block
number; this is useful because the sequence number should correspond
directly to the block number, which makes sequencing easier. This also
means we no longer need to fix the sequence number in the log on
secondary nodes because we know the block's sequence number when it's
received.
- Set the view number of a BlockNew message to whatever view the node
is currently in and don't check the view number of a BlockNew when
checking for a match. We don't care about the view number of a BlockNew
message; it doesn't make sense to associate blocks with a particular
view, since they exist outside the PBFT "realm".
- When receiving and validating a PrePrepare, all nodes now validate
that the PrePrepare matches the original BlockNew; previously this
wasn't done by the secondary nodes.

Signed-off-by: Logan Seeley <seeley@bitwise.io>