Fix PrePrepare checks
Updates how PrePrepares are checked in the following ways:
- Verify that PrePrepare message is from the primary (adds an additional
PbftError type, `NotFromPrimary`)
- Only check that a matching BlockNew message exists (same block, same
sequence number); we do not need to require that only one BlockNew
exists for this sequence number, since it's possible that more than one
can exist (i.e. if a block is determined to be invalid, a new one will
be published with the same sequence number resulting in two BlockNew
messages with the same sequence number). We only check this to make sure
that the validator has already received the block we want to perform
consensus on.
- Check if an existing PrePrepare message exists with the same sequence
number and view number but a different block; we do not need to report
an error if a PrePrepare already exists with the same sequence, view,
and block since this is just a duplicate message.
- Remove view number check in `handlers::pre_prepare` because this is
already being done in `message_log::add_message`.
- Move message adding and phase switching to `handlers::pre_prepare`
from `node::on_peer_message` for consistency and simplicity.

Signed-off-by: Logan Seeley <seeley@bitwise.io>