Remove Checkpointing
Removes the checkpointing procedure and all of its associated code. This
is no longer needed with the consensus seal and the guarantees that are
provided by the state management of the blockchain. Because the
blockchain stores all state history and each block is verified by a
consensus seal that consists of PBFT messages that prove the block
commit, we are never truly garbage collecting; the blockchain will
always have these messages and state changes. This makes checkpointing
unnecessary.

The following is a summary of the changes that are made by this shift:
- When a seal is verified, it's stored in the log for potential use in a
view change message
- If a view change is required, the seal for the last committed block is
included in the view change message as the equivalent to the proof of
the last stable checkpoint.
- Consensus seals are garbage collected after each block is committed.
Once a block commits we only keep the last valid consensus seal, since
that's all we need to provide for a view change message.
- Garbage collection now happens once the number of messages in the log
is greater than or equal to the value specified in the `max_log_size`
configuration variable.

Signed-off-by: Logan Seeley <seeley@bitwise.io>