Remove hinting
Removes the hinting mechanism since it prevents the node from adding
valid messages to the log unnecessarily; as long as a message is valid
and follows the rules necessary for being added to the log, we do not
need to backlog messages until we are at the right sequence number or
phase.

However, we need to have safeguards so that we only switch phases as a
result of a peer message if the message is for the current sequence
number.

We still need to use the backlog for PrePrepares that arrive before
their corresponding BlockNew messages, since we can't properly perform
consensus until we know the validator has that block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>