Simplify check_prepared and check_committable
Simplifies the message log's `check_prepared` and `check_committable`
methods; we don't need to check that a matching BlockNew exists for a
PrePrepare message in `check_prepared` because this is already a
requirement for the PrePrepare message to be added in the first place.
These methods can also just return `bool`s instead of `Result`s.

Signed-off-by: Logan Seeley <seeley@bitwise.io>