Add BlockNews immediately, remove NotStarted phase
We do not need to keep a backlog of BlockNew messages and wait to add
them; they can be added to the regular message log immediately when they
are recevied. Because we do not process PrePrepare messages until we
have a corresponding BlockNew, we can consider having the PrePrepare +
BlockNew as starting the algorithm. This also obviates the need for the
NotStarted phase, so it is removed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>