Remove water marking
Removes the water marking component of PBFT since it is not necessary.
This mechanism was designed to prevent a primary for selecting a
sequence number that was so high it exhausted the space of sequence
numbers. Because the primary does not decide the sequence number (it is
determined by the block number), it can't violate this anyway.

Signed-off-by: Logan Seeley <seeley@bitwise.io>