Update state.working_block
Updates the way a node tracks its working block. Since the sequence
number is strictly tied to the block number, we do not need a
"tentative" working block; we can just track whether we have a working
block or not. This removes the WorkingBlockOption struct and replaces it
with a `Option<PbftBlock>`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>