Update `on_block_commit`
Updates the `on_block_commit` method of the node in the following ways:
- No longer returns a value; previously it returned a Result, but always
`Ok(())`.
- Check that the block this BlockCommit message is for is the working
block in addition to the node's phase being `Finished`; this allows the
node to be sure it's the notification for the right block before
changing its state.
- If the node already has a BlockNew for the next block, set it as the
working block right away; this will be used by the catch-up procedure so
it's ready to catch up immediately when the next BlockNew is received.
- Primary waits to initialize block until after the node checks for a
forced view change; this prevents a node from initializing a new block,
only to find out it should no longer the primary because of a forced
view change.

Signed-off-by: Logan Seeley <seeley@bitwise.io>