Update `on_block_new` and catch-up procedure
Updates the `on_block_new` method in the following ways:
- Rearranged and commented for more logical ordering.
- No longer checks for genesis block; this is unnecessary because we're
already checking for and discarding messages for old blocks, and we
always start with sequence number >= 1.
- Always add the BlockNew message to the log as long as it's not old and
its seal is valid; previously, this was handled separated here and for
catch-up, but it can just be done here in either case.
- Check if this block can be used for catch-up; previously the catch up
method would always be called and it would decide if the block was good
for catch-up, but we only need to call that method if this is the right
block for it.
- Check if this block is the block for the current sequence number, and
if it is then update the state (working block and timeouts) and
broadcast a PrePrepare if the node is the primary.

Updates catch-up method in the following ways:
- Renames the method to `catchup` instead of `try_catchup`, since it is
now called only when catchup is possible.
- Return a Result with just a unit instead of a bool, since the
`on_block_new` method already determines if this block should be used
for catch-up.
- No longer check if block number is < 2, since this method is now only
given blocks >= 2.
- Remove "catch-up from no block new" since we should never have this
condition for catch-up.
- Only update the view if it is greater than the current view; this
prevents us from accidentally going backwards.
- Update the phase to `Committing` before committing the block so the
phase can be properly switched to `Finished`.
- Remove call to `msg_log.add_message` since this is now handled by
`on_block_new` before the `catchup` method is called.
- Instead of modifying the state here, we call `on_block_commit`
manually to update the state of the node to reflect that this block was
committed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>