Replace idle and commit timeouts with one timeout
Replace the `idle_timeout` and the `commit_timeout` with a single
`faulty_primary_timeout`; this is a departure from the PBFT definition
and the view change RFC
(https://github.com/hyperledger/sawtooth-rfcs/pull/29/files).

The new timeout is roughly equivalent to the `idle_timeout`, but with
the key difference that it is stopped when a node has both a BlockNew
and a corresponding PrePrepare for the next block rather than just a
BlockNew.

This timeout will guarantee liveness in the presence of a faulty primary
by ensuring that the primary cannot indefinitely halt the network by
failing to produce a block or a PrePrepare. However, we do not need to
use a full commit timeout to ensure that a block is actually committed
on time, because whether or not a block is committed does not depend on
the primary. Once the primary has produced a block and a PrePrepare, the
network as a whole is responsible for validating the block and sending
the required messages to actually commit it; a faulty primary cannot
prevent this on its own. This means that if there is a failure to commit
a block, then there are more than f faulty nodes, and therefore a view
change will not be able to provide liveness.

This change provides the same liveness guarantees as before, but
prevents the network from unnecessarily starting a view change when it
will not help the network continue progress.

Signed-off-by: Logan Seeley <seeley@bitwise.io>