Add NewView message to trigger view changes
Add the `NewView` message type for triggering view changes. Previously,
a view change occured simply when a node received 2f + 1 ViewChange
messages; according to the PBFT algorithm, however, the primary for the
new view should send a NewView message to the rest of the network when
it receives 2f + 1 ViewChange messages, and all nodes should wait for
this NewView message before updating their view.

Signed-off-by: Logan Seeley <seeley@bitwise.io>