Remove ViewChange protobuf, use PbftMessage
Remove the ViewChange protobuf message, and instead use the PbftMessage
protobuf for view change messages. We do not need to send a seal with a
view change message; in the original PBFT definition, the proof of the
last checkpoint (equivalent to a seal for Sawtooth PBFT) was necessary
to help all nodes catch up to a common checkpoint when the view
changed. Since we have a dedicated catch up procedure, we don't need to
do this during a view change. This means that we don't need to keep the
latest consensus seal or include it in the ViewChange message, and that
a ViewChange message can be the same a regular PbftMessage except that
it has no block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>