Implement view change timeout
Implement a view change timeout that helps guarantee liveness in the
presence of message delays. This mechanism was defined in the PBFT
paper.

When a node in view v starts a view change to view v + 1, it starts a
timer (`view_change_timeout`) for some time T (`view_change_duration`).
If the timer expires before the node receives a `NewView` message from
the new primary, the node will start a view change to view v + 2 with a
timer of 2T; if the second timer, it starts changing to v + 3 with a 3T
timer; etc.

Summary of changes:
- Duration (value of T) is added to the configuration of PBFT
- Duration and timer are added to the PBFT state
- The view the node is attempting to change to is stored in the node's
mode when view changing
- Start, stop, and check the timer at the appropriate times
- View change proposal is generalized; i.e. a node at view v doesn't
assume it is changing to view v + 1, but rather some view v' > v.

Signed-off-by: Logan Seeley <seeley@bitwise.io>