Refactor Commit handler
Place all `Commit`-specific handling into a single method
`PbftNode::handle_commit` and call this method from `on_peer_message`
when a `Commit` is received. Previously, this code was split between the
`handlers::commit` function, the message log, and the body of
`on_peer_message`.

Also makes a minor modification to the `catchup` method to account for
the removal of `handlers::commit`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>