Refactor vote verification behind single function
Move vote verification to the `PbftNode::verify_vote` function;
previously, there were separate methods for verifying a consensus seal
vote or a view change vote, but most of the verification is the same
between the two. For type-specific verification, a closure called
`verification_criteria` is passed into the `verify_vote` function.

Signed-off-by: Logan Seeley <seeley@bitwise.io>