Remove `PbftLog::get_enough_messages`
Remove the `get_enough_messages` from PbftLog and move the functionality
to the `PbftNode::build_seal` method. This is the only place it is used,
so there is no reason for it to live elsewhere.

Also makes some minor modifications to how the process works and updates
the comments accordingly.

Signed-off-by: Logan Seeley <seeley@bitwise.io>