Update backlog
Update the backlog to be explicitly PrePrepare-specific; this makes the
purpose of the backlog more clear, since it is now only used for
PrePrepares. Also removes the backlog timer in the engine's main loop,
since it provides an unnecessary delay given the timeout is used for
receiving updates earlier in the loop.

Signed-off-by: Logan Seeley <seeley@bitwise.io>