Remove type check in _broadcast_pbft_message
Remove the message type check at the beginning of
`PbftNode::_broadcast_pbft_message`; when this method is called, the
node is already ensuring that it is in the right phase to be sending the
message, so this extra check is unnecessary.

Also removes the `PbftMessageType::is_multicast` and
`PbftState::check_msg_type` methods, since this was the only place they
were used.

Signed-off-by: Logan Seeley <seeley@bitwise.io>