Remove PbftError::Timeout
This error type is unnecessary; it is only used when there is a timeout
waiting for a message from the validator, which isn't an error and can
just be ignored.

Without this error, the `handle_pbft_result` function in `engine.rs`
simply logs the error and the `handle_pbft_result` function in `node.rs`
panics with the error (since it's used for unit testing), so they are
renamed to more descriptive methods: `log_any_error` and
`panic_with_error` respectively.

Signed-off-by: Logan Seeley <seeley@bitwise.io>