Simplify message parsing in from_peer_message
Simplify the message parsing in `ParsedMessage::from_peer_message` by
directly checking the type of the message to determine which type of
message to parse it as.

Requires a small modification to `ParsedMessage::from_bytes` such that
it takes the message type and sets it for the parsed message; this is
used when a node sends a message to itself.

Signed-off-by: Logan Seeley <seeley@bitwise.io>