Remove MismatchedBlocks and NoWorkingBlock errors
Remove the `MismatchedBlocks` and `NoWorkingBlock` error types from the
`PbftError` enum in favor of using the general `InternalError` type.
These types are each used in only one place, and they do not need to be
handled in a special way, so `InternalError` is sufficient.

This commit also removes unnecessary `error!` calls where the
`MismatchedBlocks` and `NoWorkingBlock` errors were used previously,
since the `InternalError`s that are being reported now will be
propogated up the stack and logged at the top level.

Signed-off-by: Logan Seeley <seeley@bitwise.io>