Refactor Finished phase fo Finishing(block_id)
Rename the `Finished` phase to `Finishing` and store the `BlockId` of
the block the node is waiting on a BlockCommit message for. This allows
us to properly ignore unwanted BlockCommit messages without the
working_block, which is being phased out; it also helps with checking if
a block has already been committed for catch-up purposes.

Signed-off-by: Logan Seeley <seeley@bitwise.io>