Assume node has previous block for catch up
In the `catchup` procedure, do not check if the node already has the
previous block; this was done using the working_block, but is actually
unnecessary because in order to validate the consensus seal in the first
place, we are assuming the node already have the previous block since
it's needed for verifying the origin of votes against membership in the
network (need list of members in `sawtooth.consensus.pbft.peers`
setting for block the seal is for).

This assumption allows us to now catch-up to any future block as long as
the block hasn't already been committed, since we know the node already
has the the previous block whenever we get this BlockNew message.

Signed-off-by: Logan Seeley <seeley@bitwise.io>