Always initialize new block when becoming primary
When a node becomes the new primary as the result of a ViewChange,
always initialize a new block. Previously, the new node would re-use an
existing block if a valid one existed; this is problematic though,
because even if there is a valid existing block, there is no reliable
way to guarantee that the block came from the node that was the primary
at the time or that the generated block is fairly ordered.

Signed-off-by: Logan Seeley <seeley@bitwise.io>