Add InvalidMessage error type
Use the `InvalidMessage` error type when an invalid message is detected.
This error is not strictly for peer messages; it is used for the
consensus seal as well, since the seal is treated like a message. The
error type contains an arbitrary description for what kind of message
was deemed invalid and why.

Signed-off-by: Logan Seeley <seeley@bitwise.io>