Don't start view change because of invalid seal
Previously, when an invalid seal was encountered, the node would start a
view change because it deemed the primary faulty. However, if a
non-primary were to publish a block with a faulty seal, they could
arbitrarily trigger a view change even if the primary isn't faulty.
Since there is no reliable way to guarantee that a block was generated
by a legitimate primary, nodes will no longer start a view change
because of an invalid seal, and instead will simply ignore the block.
The faulty primary timeout still guarantees livenes in the case that a
primary fails to produce a block with a valid seal, and the block will
eventually be failed when a valid block of the same height is committed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>