Allow waiting to perform catchup after BlockCommit
If the node is waiting on confirmation that a block was committed and it
receives a future block, it will have to wait before performing catch-up
again. When the BlockCommit message is received, the node will check if
there is a block in the log that can be used to commit the next one.

Signed-off-by: Logan Seeley <seeley@bitwise.io>