Remove PbftBlock protobuf
This commit removes the `PbftBlock` protobuf in favor of:
- Only including the `BlockId` in a `PbftMessage`. The block's ID is the
only piece of information that's needed to correlate a message with a
specific block; furthermore, by the reducing the size of each message
that's broadcasted on the network, there will be lower overhead with
transport and (de)serialization.
- Storing a `Block` (defined by the Sawtooth Rust SDK) in its own
`HashSet` instead of in the main message log as a type of `PbftMessage`.
This eliminates the duplication provided by the `PbftBlock` (which only
really existed to reduce the size of the payload of each `PbftMessage`);
it also provides a more logical separation in the log of PBFT messages
and blocks (which are a blockchain-specific entity, and for which
concepts like "view number" don't have any meaning).

In service of the above, the following changes are made:
- Remove the `PbftBlock` protobuf and its associated implementations.
- Remove the `BlockNew` type for PBFT messages.
- Instead of adding blocks to and reading them from `PbftMessage`s, add
and read only the block ID.
- Where more information than just a block's ID is needed, use the
`Block` struct.

Signed-off-by: Logan Seeley <seeley@bitwise.io>