Fix PBFT glossary format
Use the sphinx "glossary" directive instead of an RST table,
which doesn't display properly in the HTML output.

This change allows the PBFT docs to use the "term" directive
to link to a glossary term in the future.

Note: This PR simply changes the formatting. There are no changes to the
terms or definitions.

Signed-off-by: Anne Chenette <chenette@bitwise.io>