Generalize consensus seal
Generalize the consensus seal to not be specific to the block it is put
into. This removes the unnecessary restriction that a given seal can
(and must) be placed into one and only one block: the block immediately
following the block the seal verifies. A consensus seal does not need to
be placed in a block at all. In general, the seal for block n will be
placed in block n + 1; but we want to have the flexibility of sending a
seal on its own, not as part of a block.

This change involves removing the `summary` field from the `PbftSeal`
protobuf, renaming the `previous_id` field to the more general
`block_id`, and renaming the `previous_commit_votes` field to simply
`commit_votes`.

Signed-off-by: Logan Seeley <seeley@bitwise.io>