Refactor consensus seal verification
Refactor the `PbftNode::verify_consensus_seal` method into two methods:
`verify_consensus_seal_from_block` and `verify_consensus_seal`. The
first method will be called when verifying a seal in a Block (parses the
seal from the block's payload, verifies that the seal is for the
previous block, etc.), while the 2nd verifies the contents of the seal
(validates votes). This separation allows for checking the validity of a
consensus seal that is not received as part of a block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>