Update catchup method to take seal, not block
Update the `PbftNode::catchup` method to take the consensus seal that
it will use to commit the next block rather than the block that contains
the seal. This allows for more flexibility: we can commit a block using
catchup even when the seal isn't in a block.

To simplify getting the seal from a block, the
`verify_consensus_seal_from_block` method now returns the seal in its
`Result` rather than just a unit.

Signed-off-by: Logan Seeley <seeley@bitwise.io>