Implement final block catchup procedure
Implement a procedure to commit the final block in the chain when
catching up. The standard catchup procedure does not work for this,
because it requires the seal in the next block; when the node needs to
commit the last block in the chain, there is no next block yet. The
procedure works as follows:
- When a node has caught up to the 2nd-to-last block, it broadcasts a
`SealRequest` message to the network.
- The up-to-date nodes receive the `SealRequest`:
  - If the receiving node has already committed the block, it builds a
  seal and sends it back in a `SealResponse` message.
  - If the receiving node has not yet committed the block, it adds the
  request to its log and builds/sends the seal once the block is
  committed (in `on_block_commit`).
- When the requesting node receives a `SealResponse` (it only needs one
and will ignore the others), it validates the seal and uses it to commit
the final block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>