Allow switching to Preparing after commit
This commit fixes a bug where, when a new block and its PrePrepare
message are received before the previous block gets committed, the node
will never switch to Preparing that block. To fix this, the node will
now try to switch to Preparing when one of three things happens:
- A PrePrepare is received; if the PrePrepare is for the current
sequence number and the node already has the block the message is for,
it can switch to Preparing
- A new block is received; if the block is for the current sequence
number and the node already has the PrePrepare for the block, it can
switch to Preparing.
- A block was just committed; if the node already has a block for its
next sequence number and a PrePrepare for that block, it can switch to
Preparing.

Signed-off-by: Logan Seeley <seeley@bitwise.io>