Check all messages in seal are for same block
When building a consensus seal, we must ensure that all votes that are
in the seal are for the same block, not just the same view and sequence
number. To fix this, we now group by a (block_id, view) pair instead of
just view, and count the number of messages for each block/view pair.

Signed-off-by: Logan Seeley <seeley@bitwise.io>