Verify seal votes using peers from previous block
To verify the consensus seal, we need to make sure that the votes in the
seal are all from nodes that were members of the network at the time the
block was voted on. Specifically, this means that all voters should be
in the on-chain list of peers in the block prior to the one the seal
verifies.

Before this commit, the list of voters was being taken from the block
the seal verifies itself, which is incorrect. To fix this, we add the
`previous_id` field to the consensus seal when it is built, verify that
the previous ID matches the previous ID of the block when validating the
seal, and get the list of peers using the previous ID.

Signed-off-by: Logan Seeley <seeley@bitwise.io>