Ensure liveness by reinstating commit timeout
This commit essentially reverts some of the changes in 65466d. The
`commit_timeout`, which ensures the a block gets committed within a
specified time, allows the network to get itself "unstuck" if something
goes wrong during the commit process. This differs slightly from the
`commit_timeout` that was removed in commit 65466d; specifically, the
timer is stopped when the node decides to commit the block rather than
after the block has been committed and the engine receives the
notificaiton from the validator.

Signed-off-by: Logan Seeley <seeley@bitwise.io>