Update config loading
Update the config loading procedure such that the config is created in
`main()`, passed into the `PbftEngine`, and updated with the on-chain
settings in `PbftEngine::start` using the `PbftConfig::load_settings`.

This will enable modifying the configuration using outside sources other
than on-chain settings (such as command-line arguments or a config
file).

Signed-off-by: Logan Seeley <seeley@bitwise.io>