Fix bug causing node to get stuck in ViewChanging
This fixes a bug where a node would get stuck in the `ViewChanging`
mode. The node would be able to continue to commit blocks using
catch-up, but would ignore all normal PBFT messages since it was in the
`ViewChanging` mode. This is fixed by always making sure the mode is set
to `Normal` when a block gets committed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>