Re-implement guarded PrePrepare broadcasting
This re-implements commit 52a5f3, which was inadvertently reverted.

Before broadcasting a PrePrepare to the network when a BlockNew for the
current sequence number is received, the primary must check to make sure
that the block was the one it just created (the `signer_id` of the block
should match the node's own ID). This prevents the primary from
incorrectly broadcasting a PrePrepare if another node were to publish a
block.

Signed-off-by: Logan Seeley <seeley@bitwise.io>