Replace has_block with get_block_with_id
Replace the `PbftLog::has_block` method with the `get_block_with_id`
method. This new method is more versatile, since it can provide the
same functionality as the old one with a call to `.is_some()` or
`.is_none()`, but also allows for getting the block itself (which will
be used to simplify places where `get_blocks` is called with additional
filters and checks).

Signed-off-by: Logan Seeley <seeley@bitwise.io>