Fix double-broadcasting bug for NewViews
Fixes a bug where NewView messages would be broadcasted twice; this
happened when a new primary of a 4 node network started a view change
early when it received f + 1 ViewChange messages. Because it's a 4 node
network, f=1, so f + 1 == 2f. This means that it would self-send a
ViewChange message which would trigger a NewView broadcast, then trigger
a NewView broadcast as a result of the original message that made the
count f + 1.

This is fixed by simply returning when a view change is started early;
the NewView message will still be triggered (if necessary) by the
self-sent message without sending it twice.

Signed-off-by: Logan Seeley <seeley@bitwise.io>