Fix state reset on commit
This commit fixes a bug where a node wouldn't be able to complete the
final block catch-up procedure because it didn't update it's phase
before requesting the seal (it would still be in the Finishing phase,
but when it receives the requested seal it checks if its in the
Finishing phase to make sure it doesn't commit the block twice).

Also removes the `PbftState::reset_to_start` method since it is very
simple and only used in one place now.

Signed-off-by: Logan Seeley <seeley@bitwise.io>