Only add seal request to log if for next block
Updates the seal request handler to only add the seal to the log if its
for the next block the node will commit (i.e. the message's sequence
number is the same as the node's sequence number). If the node is
already more than a block ahead, it won't be able to build the seal and
the requesting node should be able to commit its desired block using the
regular catch-up procedure. If the node is also behind the last block
committed by the network, some other node will have to build the seal
and send it to the requester, since this node also needs to catch up.

Signed-off-by: Logan Seeley <seeley@bitwise.io>