Don't verify seal when catching up after commit
When a block is committed, the node will check to see if its log
contains a block with a consensus seal that can be used to commit the
next block. It doesn't need to verify the seal here, since the seal is
verified as a prerequisite for the block to be added to the log in the
first place. Instead, it should simply parse the seal from the block and
use it for the next iteration of catch-up.

Signed-off-by: Logan Seeley <seeley@bitwise.io>