Remove old condition for checking prev block
When the node receives a new block, it must verify that it has the
previous block since it needs it to verify the consensus seal.
Previously, we checked this by seeing if: 1) the block was in the log,
or 2) the block was for the current sequence number. (2) was necessary
because if a node started up with a non-genesis chain head, it didn't
have previous block in its local log. However, this was changed in
commit 5c886cd; since that commit, the node will receive the chain head
on startup and add it to its log. So the node will always have the
previous block in its log.

Signed-off-by: Logan Seeley <seeley@bitwise.io>