Update broadcast_message to take ParsedMessage
The `PbftNode::broadcast_message` method now takes a `ParsedMessage`
instead of just the parsed bytes of the message to be broadcasted. This
serves cuts out the extra step of encoding the message to bytes, then
parsing a message from the bytes again (creating a ParsedMessage
includes serializing the message to bytes).

This commit also updates the `ParsedMessage`'s `from_pbft_message` and
`from_new_view_message` methods to set `from_self` to true; this is
because these are only used for creating `ParsedMessage`s for
self-sending.

Signed-off-by: Logan Seeley <seeley@bitwise.io>