Remove BlockCommit ignoring
This is a holdover from a time when the catch-up procedure didn't wait
for a BlockCommit message before committing the next block. Now, the
node doesn't do anything else until it receives confirmation from the
validator that the requested block was committed.

As a result of this change, the BlockId stored in the `Finishing` phase
is no longer needed, so it is also removed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>