Do not view change on invalid NewView msg
Updates the NewView handler to not view change when an invalid NewView
message is received. This is changed because it results in having to
reject potentially valid NewView messsages to prevent non-primary nodes
from arbitrarily triggerring a view change. That adds unnecessary
complexity with very little added value.

Since we no longer need to verify that the NewView is expected, we will
need to check that it is for a future view to prevent re-use of an old
NewView message.

As a result of this change, the `NotFromPrimary` error type is no longer
necessary, so it is removed.

Signed-off-by: Logan Seeley <seeley@bitwise.io>