Remove view checking from PbftLog
Remove view checking as a condition for adding messages to the log. Only
PrePrepare, Prepare, and Commit messages must be checked that they are
for the current view as a condition for acceptance into the log. Rather
than checking this for all messages types and exposing state information
to the log, the node should just enforce the view number for the
appropriate message types when they are received.

Signed-off-by: Logan Seeley <seeley@bitwise.io>