Remove previous_id from consensus seal
The `previous_id` field was added to the `PbftSeal` protobuf in e7da88d
as a way to get the list of peers from the block that represents the
state of the network at the time the block was voted on. However,
including this in the seal itself is unnecessary and redundant: to
verify that the previous ID of the seal is valid, we need to get the
block the seal proves and compare the previous IDs; instead, we can just
use the previous ID from the block itself which allows us to skip the
comparison and keep the consensus seal protobufs smaller.

Signed-off-by: Logan Seeley <seeley@bitwise.io>