Fix Prepare voting
Fixes the Prepare voting by not allowing the primary to broadcast a
Prepare message; the primary's PrePrepare message counts as its "vote"
for the Preparing phase. If the primary were allowed to broadcast a
Prepare, it could cause nodes to diverge on the blocks they commit for a
given sequence number when the network is a certain size.

Now, a non-faulty primary will not broadcast a Prepare message, and
non-faulty secondaries will not accept Prepare messages from the
primary.

Signed-off-by: Logan Seeley <seeley@bitwise.io>